+++ examples/solid/start-bun/src/routes/__root.tsx
@@ -2,7 +2,6 @@ import { TanStackDevtools } from '@tanstack/solid-devtools'
 import { HeadContent, Scripts, createRootRoute } from '@tanstack/solid-router'
 import { TanStackRouterDevtoolsPanel } from '@tanstack/solid-router-devtools'
 
-import { HydrationScript } from '@solidjs/web'
 import Header from '../components/Header'
 
 import appCss from '../styles.css?url'
@@ -37,10 +36,9 @@ function RootDocument({ children }: { children: JSX.Element }) {
   return (
     <html lang="en">
       <head>
-        <HydrationScript />
+        <HeadContent />
       </head>
       <body>
-        <HeadContent />
         <Header />
         {children}
         <TanStackDevtools