+++ examples/solid/start-counter/src/routes/__root.tsx
@@ -5,7 +5,6 @@ import {
   Scripts,
   createRootRoute,
 } from '@tanstack/solid-router'
-import { HydrationScript } from '@solidjs/web'
 import { TanStackRouterDevtools } from '@tanstack/solid-router-devtools'
 import type * as Solid from 'solid-js'
 import appCss from '~/styles/app.css?url'
@@ -21,10 +20,9 @@ function RootDocument(props: { children: Solid.JSX.Element }) {
   return (
     <html>
       <head>
-        <HydrationScript />
+        <HeadContent />
       </head>
       <body>
-        <HeadContent />
         <div>
           <Link to="/">Index</Link>
           <Link to="/about">About</Link>